+++ src/common/enums.ts
@@ -13,13 +13,6 @@ export enum UserIdentifierType {
   USER_ID = 'USER_ID',
 }
 export enum PDAStatus {
-  'Expired' = 'Expired',
-  'Revoked' = 'Revoked',
-  'Suspended' = 'Suspended',
-  'Valid' = 'Valid',
-}
-
-export enum PDAStatusV3 {
   'EXPIRED' = 'EXPIRED',
   'PENDING' = 'PENDING',
   'REVOKED' = 'REVOKED',
