+++ src/services/crypto-service.ts
@@ -1,0 +1,138 @@
+import forge from 'node-forge';
+import canonicalize from 'canonicalize';
+import { ethers } from 'ethers';
+import { createHash } from 'crypto';
+import {
+  AESCIPHERSTRATERGY,
+  ENCODING,
+  ENCRYPTIONSCHEME,
+} from '../common/constants';
+import { EncryptedAESCipher } from '../common/enums';
+
+export const hashMethod = forge.md.sha256;
+
+export class CryptoService {
+  /* The `jsonEncoder` function is a method defined within the `CryptoService` class. It takes an
+  `object` of type `any` as a parameter. */
+  jsonEncoder = (object: any) => {
+    return createHash('sha256')
+      .update(canonicalize(object) as string)
+      .digest('hex');
+  };
+
+  /**
+   * Generates a did based on Gateway V3 Protocol uses users public key
+   */
+  generateDID = (publicKey: string, domain?: string) => {
+    const sha256 = forge.md.sha256.create();
+    sha256.update(publicKey, 'utf8');
+    if (domain) return `did:gatewayId:${domain}:${sha256.digest().toHex()}`;
+    return `did:gatewayId:${sha256.digest().toHex()}`;
+  };
+
+  generateNewEtherumWallet = (privateKey: string) => {
+    return new ethers.Wallet(privateKey);
+  };
+
+  signMessage = async (
+    wallet: ethers.Wallet,
+    message: string,
+  ): Promise<string> => {
+    return await wallet.signMessage(message);
+  };
+
+  generateRSAKeyPair = () => {
+    const keypair = forge.pki.rsa.generateKeyPair({ bits: 4096 });
+    const publicPem = forge.pki.publicKeyToPem(keypair.publicKey);
+    const privatePem = forge.pki.privateKeyToPem(keypair.privateKey);
+
+    return {
+      privateKey: forge.util.encode64(privatePem),
+      publicPem: forge.util.encode64(publicPem),
+    };
+  };
+
+  sharedEncryptWithPKI = (
+    data: string,
+    ...accesses: { did: string; publicPem: string }[]
+  ) => {
+    const { cipher, key, iv } = this.generateAESCipher();
+
+    cipher.update(
+      forge.util.createBuffer(Buffer.from(data).toString(ENCODING)),
+    );
+    cipher.finish();
+
+    const aesBlob = cipher.output.getBytes();
+    const tag = cipher.mode.tag.getBytes();
+
+    const keyBlobs: Record<string, string> = {};
+    accesses.map((access) => {
+      if (access?.did && access?.publicPem) {
+        const publicPem = access.publicPem;
+        const issuePubKey = forge.pki.publicKeyFromPem(
+          forge.util.decode64(publicPem),
+        );
+        const keyBlob = issuePubKey.encrypt(key, ENCRYPTIONSCHEME, {
+          md: hashMethod.create(),
+        });
+
+        keyBlobs[access.did] = forge.util.encode64(keyBlob);
+      }
+    });
+
+    return {
+      aesBlob: forge.util.encode64(aesBlob),
+      keyBlobs,
+      iv: forge.util.encode64(iv),
+      tag: forge.util.encode64(tag),
+    };
+  };
+
+  encryptWithPKI = (data: string, publicPem: string, did: string) => {
+    return this.sharedEncryptWithPKI(data, {
+      did,
+      publicPem,
+    });
+  };
+
+  generateAESCipher = () => {
+    const key = forge.random.getBytesSync(32);
+    const iv = forge.random.getBytesSync(12);
+
+    const cipher = forge.cipher.createCipher(AESCIPHERSTRATERGY, key);
+    cipher.start({ iv });
+
+    return { cipher, key, iv };
+  };
+
+  decryptWithPKI = async (
+    cipher: EncryptedAESCipher,
+    did: string,
+    privatePem: string,
+  ): Promise<string> => {
+    const aesBlob = forge.util.decode64(cipher.aesBlob);
+    const keyBlob = forge.util.decode64(cipher.keyBlobs[did]);
+    const iv = forge.util.decode64(cipher.iv);
+    const tag = forge.util.decode64(cipher.tag);
+
+    const privateKey = forge.pki.privateKeyFromPem(privatePem);
+
+    const key = privateKey.decrypt(keyBlob, ENCRYPTIONSCHEME, {
+      md: hashMethod.create(),
+    });
+
+    const decipher = forge.cipher.createDecipher(AESCIPHERSTRATERGY, key);
+    decipher.start({
+      iv,
+      tag: forge.util.createBuffer(tag),
+    });
+    decipher.update(forge.util.createBuffer(aesBlob));
+
+    if (decipher.finish()) {
+      return Buffer.from(decipher.output.toString(), ENCODING).toString();
+    } else {
+      throw Error('AES-GCM decryption failed.');
+    }
+  };
+}
