+++ test/stubs/activity.stub.ts
