+++ test/stubs/proof.stub.ts
